Add 'selfsigned' package to support secure HTTPS server
The 'selfsigned' package has been included to allow for the creation of self-signed certificates, enabling the app to run a secure HTTPS server. This implementation enhances the security of data transmission, particularly with cookies, by providing a layer of encryption against potential attacks. The server is now capable of carrying out HTTPS-based interactions.