Refactor MailService class for dependency injection
The MailService class was refactored to use dependency injection for better testability and modularity. The transporter object is now injected into the constructor instead of being created in it. This allows functionality like sending emails or resetting passwords to be more easily unit tested and swapped out for different implementations.